Add native bf16 output support to int8_mm_dequant
Templates kdequant_mm_int32_fp16 and its launcher on output dtype, adds a
cdequant_mm_int32_bf16 C ABI entry point, and dispatches on dtype in the
Python backend. Fused bias now works for both fp16 and bf16 when the bias
dtype matches the output dtype.

Exposes a dtype kwarg on bitsandbytes.functional.int8_mm_dequant so callers
can request bf16 output directly instead of relying on a cast afterward.

Reverts the earlier test workaround now that int8_mm_dequant genuinely
returns bf16 when requested.

Addresses review feedback on #1985 (previously deferred as a follow-up).

+++ tests/test_functional.py
@@ -410,7 +410,7 @@ class TestLLMInt8Functional:
             A = A.view(-1, A.shape[-1])
             CA, statsA, _ = F.int8_vectorwise_quant(A)
             CB, statsB, _ = F.int8_vectorwise_quant(B)
-            output = F.int8_mm_dequant(F.int8_linear_matmul(CA, CB), statsA, statsB)
+            output = F.int8_mm_dequant(F.int8_linear_matmul(CA, CB), statsA, statsB, dtype=dtype)
             torch.testing.assert_close(C1.view(-1, C1.shape[-1]), output, atol=0.025, rtol=0.05)
 
     @pytest.mark.parametrize("device", get_available_devices())
